requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
python-telegram-bot==20.7
python-dotenv==1.0.0
pytz==2023.3
dhanhq==1.3.7
requests==2.31.0
